inscription ok

+++ src/main/java/com/ecam/picto/pictopro/repository/DossierMedicalRepository.java
@@ -7,5 +7,5 @@ import com.ecam.picto.pictopro.entity.DossierMedical;
 
 @Repository
 public interface DossierMedicalRepository extends JpaRepository<DossierMedical, Integer> {
-	Iterable<DossierMedical> findAllByProfessionnelId(Long id);
+	Iterable<DossierMedical> findAllByProfessionnelId(int id);
 }
